Cleaning up the files, removing unused properties and fields and classes.
