+++ src/Framework/Framework/Compilation/ViewCompiler/ViewCompilerConfiguration.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using DotVVM.Framework.Compilation.ControlTree.Resolved;
+using System.Collections.Generic;
+
+namespace DotVVM.Framework.Compilation
+{
+    public class ViewCompilerConfiguration
+    {
+        public List<Func<ResolvedControlTreeVisitor>> TreeVisitors { get; } = new List<Func<ResolvedControlTreeVisitor>>();
+    }
+}
